+++ src/main/java/com/hbm/entity/effect/EntityDrying.java
@@ -226,7 +226,7 @@ public class EntityDrying extends Entity implements IConstantRenderer, IChunkLoa
 		this.maxArea = i * i * 8;
 		this.xlast = -i;
 		if(i > 150)
-			this.speed = (int)(BombConfig.mk4);
+			this.speed = BombConfig.mk4;
 	}
 
 	public int getScale() {
