+++ src/main/java/com/hbm/entity/logic/EntityNukeExplosionMK4.java
@@ -169,7 +169,7 @@ public class EntityNukeExplosionMK4 extends Entity implements IChunkLoader {
 						rainDrop.posX = this.posX;
 						rainDrop.posY = this.posY;
 						rainDrop.posZ = this.posZ;
-						rainDrop.setScale((int)this.radius+16);
+						rainDrop.setScale(this.radius +16);
 						this.world.spawnEntity(rainDrop);
 						this.setDead();
 					}
@@ -178,7 +178,7 @@ public class EntityNukeExplosionMK4 extends Entity implements IChunkLoader {
 					rainDrop.posX = this.posX;
 					rainDrop.posY = this.posY;
 					rainDrop.posZ = this.posZ;
-					rainDrop.setScale((int)this.radius+16);
+					rainDrop.setScale(this.radius +16);
 					this.world.spawnEntity(rainDrop);
 					this.setDead();
 				}
@@ -260,8 +260,8 @@ public class EntityNukeExplosionMK4 extends Entity implements IChunkLoader {
 			r = 25;
 
 		EntityNukeExplosionMK4 mk4 = new EntityNukeExplosionMK4(world);
-		mk4.radius = (int) (r);
-		mk4.speed = (int) 1000*BombConfig.mk4/r;
+		mk4.radius = r;
+		mk4.speed = 1000 *BombConfig.mk4/r;
 		mk4.setPosition(x, y, z);
 		mk4.floodPlease = isWet(world, new BlockPos(x, y, z));
 		if(BombConfig.disableNuclear)
@@ -275,8 +275,8 @@ public class EntityNukeExplosionMK4 extends Entity implements IChunkLoader {
 			MainRegistry.logger.log(Level.INFO, "[NUKE] Initialized eX explosion at " + x + " / " + y + " / " + z + " with radius " + r + "!");
 
 		EntityNukeExplosionMK4 mk4 = new EntityNukeExplosionMK4(world);
-		mk4.radius = (int) (r);
-		mk4.speed = (int) 1000*BombConfig.mk4/r;
+		mk4.radius = r;
+		mk4.speed = 1000 *BombConfig.mk4/r;
 		mk4.setPosition(x, y, z);
 		mk4.floodPlease = isWet(world, new BlockPos(x, y, z));
 		if(BombConfig.disableNuclear)
@@ -290,8 +290,8 @@ public class EntityNukeExplosionMK4 extends Entity implements IChunkLoader {
 			MainRegistry.logger.log(Level.INFO, "[NUKE] Initialized nR explosion at " + x + " / " + y + " / " + z + " with radius " + r + "!");
 
 		EntityNukeExplosionMK4 mk4 = new EntityNukeExplosionMK4(world);
-		mk4.radius = (int) (r);
-		mk4.speed = (int) 1000*BombConfig.mk4/r;
+		mk4.radius = r;
+		mk4.speed = 1000 *BombConfig.mk4/r;
 		mk4.setPosition(x, y, z);
 		mk4.floodPlease = isWet(world, new BlockPos(x, y, z));
 		mk4.fallout = false;
